requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.4.1
uvicorn==0.12.2
uvloop==0.14.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0